Update README.md
ML-Assignment-4-Classification Problem
Breast-Cancer-Prediction
The objective of this assessment is to evaluate your understanding and ability to apply supervised learning techniques to a real-world dataset. Dataset used in this project is breast cancer dataset from sklearn library. Breast-Cancer-Prediction
Objective: The objective of this assessment is to evaluate your understanding and ability to apply supervised learning techniques to a real-world dataset.
Dataset: Use the breast cancer dataset available in the sklearn library.